+++ lib/view/shapes.dart
@@ -1,8 +1,8 @@
 part of boarding;
 
 drawCircle(CanvasElement canvas, num x, num y, num radius,
-           {num lineWidth: 1, String color: 'white',
-             String borderColor: 'black'}) {
+           {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
+  // x center, y center
   var context = canvas.getContext('2d');
   context
       ..lineWidth = lineWidth
@@ -16,9 +16,44 @@ drawCircle(CanvasElement canvas, num x, num y, num radius,
 }
 
 drawCircleWithinSquare(CanvasElement canvas, num x, num y, num length,
-                       {num lineWidth: 1, String color: 'white',
-                         String borderColor: 'black'}) {
-  drawCircle(canvas, x + length / 2, y + length / 2, length / 2,
+                       {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
+  // x left, y top
+  var r = length / 2;
+  drawCircle(canvas, x + r, y + r, r,
+             lineWidth: lineWidth, color: color, borderColor: borderColor);
+}
+
+drawPosition(CanvasElement canvas, num x, num y, {num lineWidth: 1, String color: 'black'}) {
+  var r = 1 / 2;
+  drawCircle(canvas, x + r, y + r, r, lineWidth: lineWidth, color: color);
+}
+
+// http://scienceprimer.com/drawing-regular-polygons-javascript-canvas
+drawPolygon(CanvasElement canvas, num x, num y, num sideLength, int sideCount, 
+            {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
+  // x center, y center
+  var context = canvas.getContext('2d');
+  context
+      ..lineWidth = lineWidth
+      ..fillStyle = color
+      ..strokeStyle = borderColor
+      ..beginPath()
+      ..moveTo (x +  sideLength * cos(0), y +  sideLength *  sin(0));          
+  for (var i = 1; i <= sideCount; i++) {
+    context.lineTo(x + sideLength * cos(i * 2 * PI / sideCount), 
+                   y + sideLength * sin(i * 2 * PI / sideCount));
+  }
+  context
+      ..closePath()
+      ..fill()
+      ..stroke();
+}
+
+drawPolygonWithinSquare(CanvasElement canvas, num x, num y, num length, num sideLength, int sideCount,
+                        {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
+  // x left, y top
+  var r = length / 2;
+  drawPolygon(canvas, x + r, y + r, sideLength, sideCount,
              lineWidth: lineWidth, color: color, borderColor: borderColor);
 }
 
@@ -27,8 +62,7 @@ drawCircleWithinSquare(CanvasElement canvas, num x, num y, num length,
 //http://www.html5canvastutorials.com/advanced/html5-canvas-save-drawing-as-an-image/
 // to do: draw a cloud within a shape (polygon, regular polygon)
 drawCloud(CanvasElement canvas, num x, num y, num width, num height,
-          {num lineWidth: 1, String color: 'white',
-            String borderColor: 'black'}) {
+          {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   context
       ..lineWidth = lineWidth
@@ -49,8 +83,7 @@ drawCloud(CanvasElement canvas, num x, num y, num width, num height,
 
 //http://stackoverflow.com/questions/2172798/how-to-draw-an-oval-in-html5-canvas
 drawEllipseWithinRect(CanvasElement canvas, num x, num y, num width, num height,
-                    {num lineWidth: 1, String color: 'white',
-                      String borderColor: 'black'}) {
+                    {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   var kappa = .5522848,
       ox = (width / 2) * kappa,  // control point offset horizontal
@@ -75,8 +108,7 @@ drawEllipseWithinRect(CanvasElement canvas, num x, num y, num width, num height,
 }
 
 drawLine(CanvasElement canvas, num x1, num y1, num x2, num y2,
-         {num lineWidth: 1, String color: 'black',
-           String borderColor: 'black'}) {
+         {num lineWidth: 1, String color: 'black', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   context
       ..lineWidth = lineWidth
@@ -92,8 +124,7 @@ drawLine(CanvasElement canvas, num x1, num y1, num x2, num y2,
 }
 
 drawOneOfLines(CanvasElement canvas, num x, num y, num x1, num y1,
-               {num lineWidth: 1, String color: 'black',
-                 String borderColor: 'black'}) {
+               {num lineWidth: 1, String color: 'black', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   context
       ..lineWidth = lineWidth
@@ -130,8 +161,7 @@ drawRect(CanvasElement canvas, num x, num y, num width, num height,
 }
 
 //http://stackoverflow.com/questions/1255512/how-to-draw-a-rounded-rectangle-on-html-canvas
-prepareRoundedRect(CanvasElement canvas, num x, num y, num width, num height,
-                   num radius) {
+prepareRoundedRect(CanvasElement canvas, num x, num y, num width, num height, num radius) {
   var r2d = PI / 180;
   //ensure that the radius isn't too large for x
   if ((width - x) - (2 * radius) < 0) {radius = (( width - x ) / 2);}
@@ -152,8 +182,8 @@ prepareRoundedRect(CanvasElement canvas, num x, num y, num width, num height,
       ..closePath();
 }
 
-drawRoundedRect(CanvasElement canvas, num x, num y, num width, num height,
-                {num radius: 10, num lineWidth: 1, String color: 'white',
+drawRoundedRect(CanvasElement canvas, num x, num y, num width, num height, 
+                {num radius: 10, num lineWidth: 1, String color: 'white', 
                   String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   prepareRoundedRect(canvas, x, y, x + width, y + height, radius);
@@ -166,8 +196,7 @@ drawRoundedRect(CanvasElement canvas, num x, num y, num width, num height,
 }
 
 drawSelectedRect(CanvasElement canvas, num x, num y, num width, num height,
-                 {num lineWidth: 1, String color: 'white',
-                   String borderColor: 'black'}) {
+                 {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
   const int sss = 8; // selection square size
   var context = canvas.getContext('2d');
   drawRect(canvas, x, y, width, height, lineWidth: lineWidth, color: color,
@@ -184,8 +213,7 @@ drawSelectedRect(CanvasElement canvas, num x, num y, num width, num height,
 }
 
 drawSquare(CanvasElement canvas, num x, num y, num length,
-           {num lineWidth: 1, String color: 'white',
-             String borderColor: 'black'}) {
+           {num lineWidth: 1, String color: 'white', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   context
       ..lineWidth = lineWidth
@@ -200,8 +228,8 @@ drawSquare(CanvasElement canvas, num x, num y, num length,
 
 //http://stackoverflow.com/questions/25837158/how-to-draw-a-star-by-using-canvas-html5
 drawStar(CanvasElement canvas, num x, num y, num radius,
-         {num innerRadius, int spikes: 5, num lineWidth: 1,
-           String color: '#ffff99', String borderColor: 'black'}) {
+         {num innerRadius, int spikes: 5, num lineWidth: 1, String color: '#ffff99', 
+           String borderColor: 'black'}) {
   // light yellow: #ffff99
   var rot = PI / 2 * 3;
   var sx = x;
@@ -236,15 +264,14 @@ drawStar(CanvasElement canvas, num x, num y, num radius,
 }
 
 drawStarWithinSquare(CanvasElement canvas, num x, num y, num length,
-                    {num innerRadius, int spikes: 5, num lineWidth: 1,
-                     color: 'white', String borderColor: 'black'}) {
+                    {num innerRadius, int spikes: 5, num lineWidth: 1, color: 'white', 
+                      String borderColor: 'black'}) {
   drawStar(canvas, x + length / 2, y + length / 2, length / 2,
            innerRadius: innerRadius, spikes: spikes, lineWidth: lineWidth,
            color: color, borderColor: borderColor);
 }
 
-List<Map<String, num>> prepareStars(CanvasElement canvas, int count,
-    {int spikes: 5}) {
+List<Map<String, num>> prepareStars(CanvasElement canvas, int count, {int spikes: 5}) {
   var stars = new List<Map>();
   for (var i = 0; i < count; i++) {
     var x = randomNum(canvas.width);
@@ -289,8 +316,7 @@ drawRandomStars(CanvasElement canvas, int count) {
 
 // maxWidth in pixels
 drawTag(CanvasElement canvas, num x, num y, num size, String text,
-       {num maxWidth, num lineWidth: 1,
-         String color: 'black', String borderColor: 'black'}) {
+       {num maxWidth, num lineWidth: 1, String color: 'black', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   context
       ..font = '${size}px sans-serif'
@@ -308,8 +334,8 @@ drawTag(CanvasElement canvas, num x, num y, num size, String text,
 }
 
 drawOneOfTags(CanvasElement canvas, num x, num y, num size, String text,
-             {num maxWidth, num lineWidth: 1,
-               String color: 'black', String borderColor: 'black'}) {
+             {num maxWidth, num lineWidth: 1, String color: 'black', 
+               String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   context
       ..font = '${size}px sans-serif'
@@ -323,8 +349,7 @@ drawOneOfTags(CanvasElement canvas, num x, num y, num size, String text,
 }
 
 drawTriangleWithinSquare(CanvasElement canvas, num x, num y, num length,
-                         {num lineWidth: 1, color: 'white',
-                           String borderColor: 'black'}) {
+                         {num lineWidth: 1, color: 'white', String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   //var height = 100 * (sqrt(3) / 2);
   context
@@ -342,7 +367,7 @@ drawTriangleWithinSquare(CanvasElement canvas, num x, num y, num length,
 }
 
 drawVehicle(CanvasElement canvas, num x, num y, num width, num height,
-            {num radius: 10, num lineWidth: 1, String color: 'white',
+            {num radius: 10, num lineWidth: 1, String color: 'white', 
               String borderColor: 'black'}) {
   var context = canvas.getContext('2d');
   drawRoundedRect(canvas, x, y, width, height, radius: radius,